Add unit test for UserService
Add application-test.properties file

+++ BoardGameEventBackend/src/main/java/com/example/BoardGameEventBackend/controller/UserController.java
@@ -4,10 +4,16 @@ import com.example.BoardGameEventBackend.dto.EventDto;
 import com.example.BoardGameEventBackend.dto.EventDtoMapper;
 import com.example.BoardGameEventBackend.dto.UserDto;
 import com.example.BoardGameEventBackend.dto.UserDtoMapper;
+import com.example.BoardGameEventBackend.exception.ForbiddenException;
 import com.example.BoardGameEventBackend.model.User;
 import com.example.BoardGameEventBackend.service.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
@@ -40,6 +46,16 @@ public class UserController {
     @PutMapping("/users/{id}")
     @PreAuthorize("hasAuthority('USER')")
     public UserDto updateUser(@PathVariable Long id, @Valid @RequestBody User user){
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        UserDetails principal = (UserDetails) authentication.getPrincipal();
+
+        User loggedUser = userService.findByUsername(principal.getUsername());
+
+        if(authentication.getAuthorities().stream().map(GrantedAuthority::getAuthority).noneMatch(role -> role.contains("ADMIN"))
+                && !loggedUser.getId().equals(id)){
+            throw new ForbiddenException();
+        }
+
         return UserDtoMapper.mapToUserDto(userService.updateUser(id, user));
     }
 
@@ -52,6 +68,15 @@ public class UserController {
     @DeleteMapping("/users/{id}")
     @PreAuthorize("hasAuthority('USER')")
     public void deleteUser(@PathVariable Long id){
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        UserDetails principal = (UserDetails) authentication.getPrincipal();
+
+        User loggedUser = userService.findByUsername(principal.getUsername());
+
+        if(authentication.getAuthorities().stream().map(GrantedAuthority::getAuthority).noneMatch(role -> role.contains("ADMIN"))
+                && !loggedUser.getId().equals(id)){
+            throw new ForbiddenException();
+        }
         userService.delete(id);
     }
